pyproject.toml
[project]
name = "kingdom-77-bot"
version = "3.9.0"
description = "Kingdom-77 Discord Bot v3.9 - Enterprise-level Discord bot with premium features, credits economy, multi-language support, and payment integration"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.1.1",
    "deep-translator>=1.11.4",
    "langdetect>=1.0.9",
    "motor>=3.3.2",
    "pymongo>=4.6.1",
    "redis>=5.0.0",
]
